tidied up directories more
Directory structure is now suitable for simply dumping everything into
one web root.
In the long run this is not desirable, instead we should have on folder
for frontend, one for backend and one for installation. However,
classical PHP/Apache architecture makes this hard to achieve. Hence,
this will have to wait until the backend is rewritten in python.