Frontend: Worked around apparent (p)react bug.
Instead of <input type="submit">, now <button type="submit"> is used.
Otherwise (p)react seems to not update the value field of the <input
type="submit"> element (which is responsible for the text on the button)
when doing a diff on the DOM tree when switching from the shoppinglist
to other input fields. As a result, a warning was logged to the console
stating that the value "Speichern" is not allowed in the number input
field with the label "amount"...