API: Fix handling of non-utf8 db data
The charset for the db connection is now set to utf8mb4 on every
connection.

Previously, the code relied on the database to have the
default charset set to utf8. This led to errors when this was not the
case, in particular when serializing data from the db as JSON, because
JSON requires utf8.