Fix wrong timestamp when generating charts
When generating the charts after a POST or DELETE request e.g. to
/backend/api/hoursofwork/entries.php, the associated timestamp is sent
to the client as JSON and, once chart generation is finished, can be
found at /generated/hoursofwork/timestamp.

The two timestamps differed because they had different numbers of
decimals. Therefore, in 50% of the cases it appeared to the client that
chart generation never finished.

Fix this by serializing the timestamp via json_encode() (instead of
strval()) in every case: when sending it to the client and when handing
it to the python script that generates the charts.